Add implementations for basic CA functions
Service can show CA cert, refresh CA, and sign CSR for a CA
mounted in vault. Uses a vault ca module in firkinize

Also added a version endpoint. This will be used by clients (in particular
the hostagent installer) to decide whether or not a controller is capable
of providing individual host certs.

testing:
Need to add unit and integration tests. Used curl to test the
service. Created a simple test program to test the methods in
vouch.