Add init-region to vouch
Add an init-region script to the vouch container image, and tweak
a few of the configs now that I've settled on a schema for vouch's
consul entries. For details on the schema, see the top of init-region.

testing: deployed into decco using the decco_pmk testbed - the deployment
code is modified in firkinize. Verified that the services behave correctly
and all the apis work. Still need to test the noauth version of the
container a bit more and add a full integration test to our suite.